Update the igatool.
The original igatool's unpack mode just extract files to the current
directory, the new igatool will create a folder and extract the file in
it. And the original igatool do not support pack 'some' of the original
files becaues it doesn't modify the filename block. The new version
support pack any folder with resource files. This will be more convenient
if we don't want pack all of the resource files because it will make a
big file, such as bgimage.iga and fgimage.iga. Also do not need
as-util.h anymore.